+++ plugins/search-backend-module-elasticsearch/src/engines/index.ts
@@ -17,7 +17,6 @@
 export {
   decodePageCursor as decodeElasticSearchPageCursor,
   ElasticSearchSearchEngine,
-  elasticSearchEngineModule,
 } from './ElasticSearchSearchEngine';
 export { isOpenSearchCompatible } from './ElasticSearchClientOptions';
 export type {
