+++ plugins/search-backend-module-elasticsearch/src/index.ts
@@ -23,7 +23,6 @@
 export {
   decodeElasticSearchPageCursor,
   ElasticSearchSearchEngine,
-  elasticSearchEngineModule,
   isOpenSearchCompatible,
 } from './engines';
 export type {
