+++ plugins/search-backend/src/index.ts
@@ -21,4 +21,3 @@
  */
 
 export * from './service/router';
-export * from './plugin';
